fix(fixed): move logic from selectors to transformators
